rename package to network
