+++ pkg/controller/ground/bootstrap.go
@@ -15,10 +15,10 @@ import (
 	openstack_project "github.com/sapcc/kubernikus/pkg/client/openstack/project"
 	"github.com/sapcc/kubernikus/pkg/controller/config"
 	"github.com/sapcc/kubernikus/pkg/controller/ground/bootstrap"
-	"github.com/sapcc/kubernikus/pkg/controller/ground/bootstrap/cni"
 	"github.com/sapcc/kubernikus/pkg/controller/ground/bootstrap/csi"
 	"github.com/sapcc/kubernikus/pkg/controller/ground/bootstrap/dns"
 	"github.com/sapcc/kubernikus/pkg/controller/ground/bootstrap/gpu"
+	"github.com/sapcc/kubernikus/pkg/controller/ground/bootstrap/network"
 	"github.com/sapcc/kubernikus/pkg/util"
 	"github.com/sapcc/kubernikus/pkg/version"
 )
@@ -109,7 +109,7 @@ func SeedKluster(clients config.Clients, factories config.Factories, images vers
 	}
 	if !kluster.Spec.NoCloud {
 		if ok, _ := util.KlusterVersionConstraint(kluster, ">= 1.24"); ok {
-			if err := cni.SeedCNIConfig(kubernetes, *kluster.Spec.ClusterCIDR, images.Versions[kluster.Spec.Version]); err != nil {
+			if err := network.SeedCNIConfig(kubernetes, *kluster.Spec.ClusterCIDR, images.Versions[kluster.Spec.Version]); err != nil {
 				return errors.Wrap(err, "seed cni config")
 			}
 		}
